trying to clean the popup menus
